feat: Lesson remaining time block
This block is only available for lesson/view screens. When added, it
will display a block that shows the required time for the lesson, and
will count down the remaining time. When meeting the time requirement,
it will say, "Complete!"

The block is not automatically displayed on all lessons. It will only be
displayed if you add it to the lesson page.

The dynamic nature of updating the remaining time is entirely handled on
the client side and is estimated. At the time of rendering the block,
the time information is written in data attributes in the block, and the
JavaScript uses it to update the block every second.

Closes #37.